source/langchain-docs/langchain/projects/04_customer_service/requirements.txt
# Core Dependencies
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.28
langchain-openai==0.2.14

# Vector Store
faiss-cpu==1.9.0.post1

# API (Optional)
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0

# Monitoring & Logging
structlog==24.4.0

# Development & Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.2
httpx==0.28.1
black==24.10.0
flake8==7.1.1

# Deployment
gunicorn==23.0.0
